int-overflow-bug

+++ server/src/org/sogive/server/DonationServlet.java
@@ -83,7 +83,7 @@ public class DonationServlet {
 		SearchResponse sr = s.get();
 		List<Map> hits = sr.getHits();
 		List<Map> hits2 = Containers.apply(h -> (Map)h.get("_source"), hits);
-		int total = sr.getTotal();
+		long total = sr.getTotal();
 		JsonResponse output = new JsonResponse(state, new ArrayMap(
 				"hits", hits2,
 				"total", total
